clean up the code

+++ controllers/workExperienceController.js
@@ -1,35 +1,50 @@
-const { WorkExperience } = require("../models");
+const { WorkExperience } = require('../models');
 
 class WorkExperienceController {
-  static findAll = async (req, res, next) => {
+  static async findWorkExperiences(req, res, next) {
     try {
-      const data = await WorkExperience.findAll();
-      res.status(200).json(data);
-    } catch (err) {
-      next(err);
-    }
-  };
+      const { id } = req.userLogged;
 
-  static findOne = async (req, res, next) => {
-    const { id } = req.params;
-    try {
-      const data = await WorkExperience.findOne({ where: { id } });
+      const data = await WorkExperience.findAll({
+        where: { user_id: id },
+      });
+
       if (data) {
         res.status(200).json(data);
       } else {
-        throw { name: "data not found" };
+        throw { name: 'ErrorNotFound' };
       }
-    } catch (err) {
-      next(err);
+    } catch (error) {
+      next(error);
     }
-  };
+  }
 
-  static create = async (req, res, next) => {
+  static async findWorkExperience(req, res, next) {
     try {
-      const { user_id, job_title, company, start_date, end_date, description, job_level, salary, salary_frequency } =
-        req.body;
+      const { id } = req.userLogged;
+      const { workExperienceId } = req.params;
+
+      const data = await WorkExperience.findOne({
+        where: { id: workExperienceId, user_id: id },
+      });
+
+      if (data) {
+        res.status(200).json(data);
+      } else {
+        throw { name: 'ErrorNotFound' };
+      }
+    } catch (error) {
+      next(error);
+    }
+  }
+
+  static async createWorkExperience(req, res, next) {
+    try {
+      const { id } = req.userLogged;
+      const { job_title, company, start_date, end_date, description, job_level, salary, salary_frequency } = req.body;
+
       const data = await WorkExperience.create({
-        user_id,
+        user_id: id,
         job_title,
         company,
         start_date,
@@ -39,40 +54,66 @@ class WorkExperienceController {
         salary,
         salary_frequency,
       });
-      res.status(201).json({ message: "New Work Experience created succesfully" });
-    } catch (err) {
-      next(err);
+      res.status(201).json({ ...data.dataValues, message: 'Successfully add work experience!' });
+    } catch (error) {
+      next(error);
     }
-  };
+  }
 
-  static update = async (req, res, next) => {
+  static async updateWorkExperience(req, res, next) {
     try {
-      const { id } = req.params;
-      const { user_id, job_title, company, start_date, end_date, description, job_level, salary, salary_frequency } =
-        req.body;
-      const data = await WorkExperience.update(
-        { user_id, job_title, company, start_date, end_date, description, job_level, salary, salary_frequency },
-        { where: { id } }
-      );
-      if (data === 0) {
-        throw { name: "not found" };
+      const { id } = req.userLogged;
+      const { workExperienceId } = req.params;
+      const { job_title, company, start_date, end_date, description, job_level, salary, salary_frequency } = req.body;
+
+      const findWorkExperience = await WorkExperience.findOne({
+        where: { id: workExperienceId, user_id: id },
+      });
+
+      if (findWorkExperience) {
+        const data = await WorkExperience.update(
+          {
+            job_title,
+            company,
+            start_date,
+            end_date,
+            description,
+            job_level,
+            salary,
+            salary_frequency,
+          },
+          { where: { id: workExperienceId, user_id: id } }
+        );
+        res.status(201).json({
+          message: 'Successfully update work experience!',
+        });
       } else {
-        res.status(200).json({ message: `Work Experience with id ${id} successfully updated` });
+        throw { name: 'ErrorNotFound' };
       }
-    } catch (err) {
-      next(err);
+    } catch (error) {
+      next(error);
     }
-  };
+  }
 
-  static destroy = async (req, res, next) => {
+  static async destroyWorkExperience(req, res, next) {
     try {
-      const { id } = req.params;
-      await WorkExperience.destroy({ where: { id } });
-      res.status(200).json({ message: `Work Experience with id ${id} success deleted` });
-    } catch (err) {
-      next(err);
+      const { id } = req.userLogged;
+      const { workExperienceId } = req.params;
+
+      const findWorkExperience = await WorkExperience.findOne({
+        where: { id: workExperienceId, user_id: id },
+      });
+
+      if (findWorkExperience) {
+        const data = await WorkExperience.destroy({ where: { id: workExperienceId, user_id: id } });
+        res.status(200).json({ message: 'Successfully delete work experience!' });
+      } else {
+        throw { name: 'ErrorNotFound' };
+      }
+    } catch (error) {
+      next(error);
     }
-  };
+  }
 }
 
 module.exports = WorkExperienceController;
